requirements.txt
# Audio I/O
pyaudio>=0.2.14
numpy>=1.24.0

# VAD (Voice Activity Detection)
onnxruntime>=1.16.0

# STT — local
faster-whisper>=1.0.0

# LLM — NVIDIA NIM (OpenAI-compatible)
openai>=1.12.0

# TTS — local Kokoro
kokoro-onnx>=0.1.0
soundfile>=0.12.1

# Sarvam AI (STT / LLM / TTS via REST API)
requests>=2.31.0

# Configuration
python-dotenv>=1.0.0

# Indic Agent — fully local, CPU-only (IndicWhisper STT + Indic Parler-TTS)
# Install PyTorch CPU build first (separate index, not included here):
#   pip install torch --index-url https://download.pytorch.org/whl/cpu
# Install Parler-TTS from GitHub:
#   pip install git+https://github.com/huggingface/parler-tts.git
torch>=2.1.0
transformers>=4.40.0
accelerate>=0.27.0
huggingface-hub>=0.22.0
